+++ examples/main/shared/viewer/ViewerPlugins.ts
@@ -180,11 +180,17 @@ const config: RichContentViewerProps['config'] = {
     resolveFileUrl: () =>
       new Promise(resolve =>
         setTimeout(
-          () => resolve('https://www.w3.org/wai/er/tests/xhtml/testfiles/resources/pdf/dummy.pdf'),
+          // eslint-disable-next-line max-len
+          () =>
+            resolve(
+              // eslint-disable-next-line max-len
+              'https://cms.education.gov.il/NR/rdonlyres/BFBDB737-89E9-4B70-A1FF-1122B7AE8F1D/69371/14_HEB_ClassRecipeBook.pdf'
+            ),
           1000
         )
       ),
     downloadTarget: '_blank',
+    adobeAPIKey: '77266adc759d4dc2a9e071e3bde25095',
   },
   [VIDEO_TYPE]: {
     getVideoUrl: src => `https://video.wixstatic.com/${src.pathname}`,