+++ examples/storybook/src/shared/resources/experimentsExample.ts
@@ -19,5 +19,6 @@ export const EXPERIMENTS = [
   { name: 'normalizePoll', scope: 'ricos' },
   { name: 'optimizeFullScreenModal', scope: 'ricos' },
   { name: 'useFilePluginAutoDownloadLinkRef', scope: 'ricos' },
+  { name: 'enableFilePluginPDFViewer', scope: 'ricos' },
   { name: 'encAutoImageUrls', scope: 'ricos' },
 ];