+++ packages/plugin-file-upload/web/src/createFileUploadPlugin.ts
@@ -8,7 +8,7 @@ import type { CreatePluginFunction } from 'wix-rich-content-common';
 
 const createFileUploadPlugin: CreatePluginFunction<FilePluginEditorConfig> = config => {
   const type = FILE_UPLOAD_TYPE;
-  const { helpers, t, [type]: settings = {}, ...rest } = config;
+  const { helpers, t, [type]: settings = {}, isMobile, experiments, ...rest } = config;
   return createBasePlugin({
     component: createBaseMediaPlugin(Component),
     type: FILE_UPLOAD_TYPE,
@@ -16,11 +16,16 @@ const createFileUploadPlugin: CreatePluginFunction<FilePluginEditorConfig> = con
       helpers,
       t,
       settings,
+
+      isMobile,
+      experiments,
     }),
     helpers,
     settings,
     t,
     defaultPluginData: DEFAULTS,
+    isMobile,
+    experiments,
     ...rest,
   });
 };