+++ packages/plugin-file-upload/web/src/file-upload-component.jsx
@@ -4,7 +4,17 @@ import FileUploadViewer from './file-upload-viewer';
 import { FILE_UPLOAD_TYPE } from './types';
 
 function FileUploadComponent(props) {
-  const { componentData, theme, setComponentUrl, t, isMobile, tempData, isLoading } = props;
+  const {
+    componentData,
+    theme,
+    setComponentUrl,
+    t,
+    isMobile,
+    tempData,
+    isLoading,
+    locale,
+    settings,
+  } = props;
   return (
     <FileUploadViewer
       componentData={componentData}
@@ -14,6 +24,8 @@ function FileUploadComponent(props) {
       t={t}
       isMobile={isMobile}
       tempDataPlaceHolder={tempData}
+      locale={locale}
+      settings={settings}
     />
   );
 }
@@ -26,6 +38,8 @@ FileUploadComponent.propTypes = {
   isMobile: PropTypes.bool,
   tempData: PropTypes.object,
   isLoading: PropTypes.bool,
+  locale: PropTypes.string,
+  settings: PropTypes.object,
 };
 
 FileUploadComponent.defaultProps = {