+++ packages/plugin-file-upload/web/src/pdfViewer/index.ts
@@ -1,0 +1,3 @@
+export { PDFViewer } from './pdfViewer';
+
+export { PDF_STATUS, PDF_OPTIONS, PDF_CUSTOMIZATIONS } from './pdfConsts';