+++ packages/plugin-file-upload/web/src/toolbar/createToolbar.ts
@@ -1,17 +1,30 @@
 import createInsertButtons from './insert-buttons';
 import createInlineButtons from './inline-buttons';
-import type { CreatePluginToolbar, TranslationFunction } from 'wix-rich-content-common';
+import type {
+  CreatePluginToolbar,
+  TranslationFunction,
+  AvailableExperiments,
+} from 'wix-rich-content-common';
 import type { FilePluginEditorConfig } from '../types';
 
 const createToolbar: CreatePluginToolbar = ({
   settings,
   t,
+  isMobile,
+  experiments,
 }: {
   t: TranslationFunction;
   settings: FilePluginEditorConfig;
+  isMobile?: boolean;
+  experiments?: AvailableExperiments;
 }) => {
   return {
-    InlineButtons: createInlineButtons({ settings, t }),
+    InlineButtons: createInlineButtons({
+      settings,
+      t,
+      isMobile,
+      experiments,
+    }),
     InsertButtons: createInsertButtons({ settings, t }),
     name: 'FileUpload',
   };