+++ website/infra/editor/config/fileMockURL.js
@@ -1,5 +1,6 @@
+/* eslint-disable max-len */
 export const FILE_URLS = {
   mp3: 'https://static.wixstatic.com/mp3/f0f74f_43f712be42dc4377a075dcad3c358a80.mp3',
-  pdf: 'https://www.w3.org/wai/er/tests/xhtml/testfiles/resources/pdf/dummy.pdf',
+  pdf: 'https://cms.education.gov.il/NR/rdonlyres/BFBDB737-89E9-4B70-A1FF-1122B7AE8F1D/69371/14_HEB_ClassRecipeBook.pdf',
   jpg: 'https://static.wixstatic.com/media/feac6b9cf32640d3ac325c905f2b3786.jpg',
 };